+++ sumologic/data_source_sumologic_http_source.go
@@ -46,6 +46,14 @@ func dataSourceSumologicHTTPSource() *schema.Resource {
 				Type:     schema.TypeString,
 				Computed: true,
 			},
+			"token": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+			"base_url": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
 		},
 	}
 
@@ -54,25 +62,46 @@ func dataSourceSumologicHTTPSource() *schema.Resource {
 func dataSourceSumologicHTTPSourceRead(d *schema.ResourceData, meta interface{}) error {
 	c := meta.(*Client)
 
-	id, _ := strconv.Atoi(d.Id())
-	var collectorId int64
+	var collectorId int
 	switch cid := d.Get("collector_id").(type) {
 	case int:
-		collectorId = int64(cid)
-	case int64:
 		collectorId = cid
+	case int64:
+		collectorId = int(cid)
 	default:
 		return fmt.Errorf("unknown data type of collector_id: %T, value: %v", cid, cid)
 	}
-	source, err := c.GetSourceName(collectorId, d.Get("name").(string))
 
-	if err != nil {
-		return err
+	var source *HTTPSource
+	var err error
+
+	// If we have both collector_id and name, look up by name first
+	if d.Get("name").(string) != "" {
+		baseSource, err := c.GetSourceName(int64(collectorId), d.Get("name").(string))
+		if err != nil {
+			return err
+		}
+		if baseSource == nil {
+			d.SetId("")
+			return fmt.Errorf("HTTP source not found")
+		}
+		// Now get the full HTTP source with all fields
+		source, err = c.GetHTTPSource(collectorId, baseSource.ID)
+		if err != nil {
+			return err
+		}
+	} else if d.Id() != "" {
+		// If we have an ID, get directly
+		id, _ := strconv.Atoi(d.Id())
+		source, err = c.GetHTTPSource(collectorId, id)
+		if err != nil {
+			return err
+		}
 	}
 
 	if source == nil {
 		d.SetId("")
-		return fmt.Errorf("HTTP source not found, removing from state: %v - %v", id, err)
+		return fmt.Errorf("HTTP source not found")
 	}
 
 	d.SetId(strconv.Itoa(source.ID))
@@ -82,6 +111,8 @@ func dataSourceSumologicHTTPSourceRead(d *schema.ResourceData, meta interface{})
 	d.Set("timezone", source.TimeZone)
 	d.Set("multiline", source.MultilineProcessingEnabled)
 	d.Set("url", source.Url)
+	d.Set("token", source.Token)
+	d.Set("base_url", source.BaseUrl)
 
 	return nil
 }