+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## X.Y.Z (Unreleased)
+
 * Add new change notes here
 
 ## 3.2.3 (January 28, 2026)
 ENHANCEMENTS:
+* Updated sumologic_http_source resource and data source:** Added `token` and `base_url` computed attributes to expose use in authentication via header
 * Added update support for `sumologic_ot_collector` resource.
 * Suppressed system-managed `sumo.disco.*` tags so they no longer appear in Terraform plans or state when reading OT Collectors.
 * Added IsEnabled field in `sumologic_source_template` resource.